--- ion/services/sa/product/test/test_data_product_management_service_meta.py
@@ -1,45 +1,0 @@
-#from pyon.ion.endpoint import ProcessRPCClient
-from pyon.public import  log, IonObject
-from pyon.util.int_test import IonIntegrationTestCase
-from ion.services.sa.product.data_product_management_service import DataProductManagementService
-from interface.services.coi.iresource_registry_service import ResourceRegistryServiceClient
-from interface.services.dm.iingestion_management_service import IngestionManagementServiceClient
-from interface.services.dm.ipubsub_management_service import PubsubManagementServiceClient
-from interface.services.sa.idata_product_management_service import  DataProductManagementServiceClient
-from interface.services.sa.idata_acquisition_management_service import DataAcquisitionManagementServiceClient
-
-from nose.plugins.attrib import attr
-
-
-from ion.services.sa.product.data_product_impl import DataProductImpl
-from ion.services.sa.resource_impl.resource_impl_metatest_integration import ResourceImplMetatestIntegration
-
-
-
-
-
-@attr('META', group='sa')
-#@unittest.skip('not working')
-class TestDataProductManagementServiceMeta(IonIntegrationTestCase):
-
-    def setUp(self):
-        # Start container
-        #print 'instantiating container'
-        self._start_container()
-
-        self.container.start_rel_from_url('res/deploy/r2deploy.yml')
-
-        print 'started services'
-
-        # Now create client to DataProductManagementService
-        self.client = DataProductManagementServiceClient(node=self.container.node)
-        self.rrclient = ResourceRegistryServiceClient(node=self.container.node)
-        self.damsclient = DataAcquisitionManagementServiceClient(node=self.container.node)
-        self.pubsubcli =  PubsubManagementServiceClient(node=self.container.node)
-        self.ingestclient = IngestionManagementServiceClient(node=self.container.node)
-
-
-rimi = ResourceImplMetatestIntegration(TestDataProductManagementServiceMeta, DataProductManagementService, log)
-rimi.test_all_in_one(True)
-
-rimi.add_resource_impl_inttests(DataProductImpl, {})
